Add TaskManager.js and drag&drop.js
* Separate all managamente logic of the task array to its own class named TaskManager
* Separate all drag&drop logic to its own file named drag&drop.js
* Refactor all relevant code
