+++ src/index.js
@@ -1,45 +1,29 @@
-import './style.css';
-import Task from './Task.js';
+import dragoverHandler from './drag&drop.js';
+import Enter from './enter-arrow.png';
 import { appendImg } from './images.js';
 import Refresh from './refresh.png';
-import Enter from './enter-arrow.png';
+import './style.css';
+import Task from './Task.js';
+import TaskManager from './TaskManager.js';
 
-const container = document.querySelector('#to-do-list');
+const taskElementsContainer = document.querySelector('#to-do-list');
 const refreshContainer = document.querySelector('#refresh');
 const enterContainer = document.querySelector('#enter');
 const addTask = document.querySelector('#add-task');
 const enter = document.querySelector('#enter');
 const warning = document.querySelector('#warning');
 const clearAll = document.querySelector('#clear-all');
-
-const dragoverHandler = (e) => {
-  e.preventDefault();
-  const y = e.clientY || e.touches[0].clientY;
-  const afterElement = Task.getDragAfterElement(container, y);
-  const draggable = document.querySelector('.dragging');
-  const draggableIndex = Task.tasks.findIndex((task) => task.node === draggable);
-  let afterElementIndex;
-  if (!afterElement) {
-    container.appendChild(draggable);
-    afterElementIndex = Task.tasks.length - 1;
-  } else {
-    container.insertBefore(draggable, afterElement);
-    afterElementIndex = Task.tasks.findIndex((task) => task.node === afterElement) - 1;
-  }
-  if (draggableIndex > afterElementIndex) afterElementIndex += 1;
-  Task.tasks[draggableIndex].afterElementIndex = afterElementIndex;
-};
+const taskManager = new TaskManager();
 
 window.onload = () => {
-  Task.loadLocalTasks();
-  Task.populateAll(container);
+  taskManager.loadLocalTasks().populateAll(taskElementsContainer);
   appendImg(Refresh, refreshContainer);
   appendImg(Enter, enterContainer);
 
   enter.addEventListener('click', () => {
     if (addTask.value) {
-      const task = new Task({ description: addTask.value });
-      task.add().populate(container, task.index - 1);
+      const task = new Task({ description: addTask.value, index: taskManager.tasks.length + 1 });
+      taskManager.add(task).populate(taskElementsContainer, task);
       addTask.value = '';
     } else {
       warning.classList.remove('d-none');
@@ -54,14 +38,14 @@ window.onload = () => {
   });
 
   clearAll.addEventListener('click', () => {
-    Task.clearAllCompleted();
+    taskManager.clearAllCompleted();
   });
 
-  container.addEventListener('dragover', (e) => {
-    dragoverHandler(e);
+  taskElementsContainer.addEventListener('dragover', (e) => {
+    dragoverHandler(e, taskManager.tasks, taskElementsContainer);
   });
 
-  container.addEventListener('touchmove', (e) => {
-    dragoverHandler(e);
+  taskElementsContainer.addEventListener('touchmove', (e) => {
+    dragoverHandler(e, taskManager.tasks, taskElementsContainer);
   });
 };
