+++ webpack.config.js
@@ -4,6 +4,7 @@ const HtmlWebpackPlugin = require('html-webpack-plugin');
 module.exports = {
   mode: 'production',
   entry: './src/index.js',
+  devtool: 'inline-source-map',
   devServer: {
     static: './dist',
   },
